Tries to unfuck winui project one more time
