+++ Material.Icons.WinUI3.Demo/MainWindow.xaml
@@ -1,60 +1,27 @@
 <?xml version="1.0" encoding="utf-8"?>
+
 <Window
-    x:Class="Material.Icons.WinUI3.Demo.MainWindow"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Material.Icons.WinUI3.Demo"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:system="using:System"
-    xmlns:winUi3="using:Material.Icons.WinUI3"
-    mc:Ignorable="d"
-    d:DataContext="{d:DesignInstance local:MainViewModel}">
+  x:Class="Material.Icons.WinUI3.Demo.MainWindow"
+  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+  xmlns:local="using:Material.Icons.WinUI3.Demo"
+  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+  xmlns:system="using:System"
+  xmlns:winUi3="using:Material.Icons.WinUI3"
+  mc:Ignorable="d"
+  d:DataContext="{d:DesignInstance local:MainViewModel}">
 
   <Grid x:Name="Root">
-    <Grid.Resources>
-      <local:StringJoinConverter x:Key="StringJoinConverter" Separator="{x:Bind system:Environment.NewLine}" />
-    </Grid.Resources>
-
-    <Grid.RowDefinitions>
-      <RowDefinition Height="*" />
-      <RowDefinition Height="Auto" />
-    </Grid.RowDefinitions>
-
-    <Grid.ColumnDefinitions>
-      <ColumnDefinition Width="*" />
-      <ColumnDefinition Width="*" />
-    </Grid.ColumnDefinitions>
-
-    <GridView
-      Grid.ColumnSpan="2"
-      ItemsSource="{Binding Kinds.View}"
-      SelectedItem="{Binding Group, Mode=TwoWay}">
-
-      <GridView.ItemTemplate>
-        <DataTemplate x:DataType="local:PackIconKindGroup">
-          <Grid Width="64" Background="Transparent">
-            <Grid.RowDefinitions>
-              <RowDefinition Height="32" />
-              <RowDefinition Height="32" />
-            </Grid.RowDefinitions>
-            <winUi3:MaterialIcon
-              Grid.Row="0"
-              Kind="{Binding Kind}"
-              Width="32"
-              Height="32" />
-            <TextBlock
-              Grid.Row="1"
-              Text="{Binding DisplayName}"
-              TextTrimming="CharacterEllipsis"
-              HorizontalAlignment="Center" />
-          </Grid>
-        </DataTemplate>
-      </GridView.ItemTemplate>
-    </GridView>
-
-    <TextBox Grid.Row="1" Grid.Column="0" Margin="8" Text="{Binding SearchText, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
-    <TextBox Grid.Row="1" Grid.Column="1" Margin="8" IsReadOnly="True" Text="{Binding CopyText}" />
+    <winUi3:MaterialIcon FontSize="10" Background="Red" />
+    <winUi3:MaterialIcon FontSize="10" Background="Red"
+                         IconSize="20" />
+    <winUi3:MaterialIcon FontSize="10" Background="Red"
+                         Width="20" Height="20" />
+    <winUi3:MaterialIcon FontSize="10" Background="Red"
+                         Width="20" Height="20" IconSize="10" />
+    <winUi3:MaterialIcon FontSize="10" Background="Red"
+                         Width="20" Height="20" IconSize="40" />
   </Grid>
 
 </Window>
