+++ Material.Icons.WinUI3.Demo/Material.Icons.WinUI3.Demo.csproj
@@ -12,14 +12,9 @@
     <UseWinUI>true</UseWinUI>
     <EnableMsixTooling>true</EnableMsixTooling>
     <Nullable>enable</Nullable>
-  </PropertyGroup>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <WindowsPackageType>None</WindowsPackageType>
 
-  <!-- Workaround, see https://github.com/microsoft/WindowsAppSDK/issues/4696 -->
-  <ItemGroup Condition="'$(NETCoreSdkVersion)' == '8.0.401'">
-    <PackageReference Include="Microsoft.Windows.CsWinRT" Version="2.1.1" />
-  </ItemGroup>
-
-  <PropertyGroup Condition="'$(NETCoreSdkVersion)' == '8.0.401'">
     <WindowsSdkPackageVersion>10.0.22621.38</WindowsSdkPackageVersion>
   </PropertyGroup>
 
@@ -37,6 +32,7 @@
     <PackageReference Include="DynamicData" Version="9.0.4" />
     <PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.26100.1" />
     <PackageReference Include="Microsoft.WindowsAppSDK" Version="1.6.240829007" />
+    <PackageReference Include="Microsoft.Windows.CsWinRT" Version="2.1.1" />
     <Manifest Include="$(ApplicationManifest)" />
   </ItemGroup>
