+++ Material.Icons.WinUI3/Themes/Generic.xaml
@@ -27,8 +27,8 @@
 
               <VisualState x:Name="UseIconSizeState">
                 <VisualState.StateTriggers>
-                  <local:ComparerTrigger ExpectedValue="{x:Bind system:Double.NaN}" IsEqual="False"
-                                         Value="{Binding IconSize, RelativeSource={RelativeSource TemplatedParent}}"/>
+                  <local:ComparerDoubleTrigger IsEqual="False"
+                                               Value="{Binding IconSize, RelativeSource={RelativeSource TemplatedParent}}" />
                 </VisualState.StateTriggers>
                 <VisualState.Setters>
                   <!-- Set Viewbox dimensions to IconSize -->
@@ -42,12 +42,11 @@
               <!-- State: Use FontSize when IconSize, Width, AND Height are ALL NaN -->
               <VisualState x:Name="UseFontSizeState">
                 <VisualState.StateTriggers>
-                  <local:ComparerTrigger ExpectedValue="{x:Bind system:Double.NaN}" 
-                                         Value="{Binding Width, RelativeSource={RelativeSource TemplatedParent}}"/>
-                  <local:ComparerTrigger ExpectedValue="{x:Bind system:Double.NaN}" 
-                                         Value="{Binding Height, RelativeSource={RelativeSource TemplatedParent}}"/>
-                  <local:ComparerTrigger ExpectedValue="{x:Bind system:Double.NaN}" 
-                                         Value="{Binding IconSize, RelativeSource={RelativeSource TemplatedParent}}"/>
+                  <local:ComparerDoubleTrigger Value="{Binding Width, RelativeSource={RelativeSource TemplatedParent}}" />
+                  <local:ComparerDoubleTrigger
+                    Value="{Binding Height, RelativeSource={RelativeSource TemplatedParent}}" />
+                  <local:ComparerDoubleTrigger
+                    Value="{Binding IconSize, RelativeSource={RelativeSource TemplatedParent}}" />
                 </VisualState.StateTriggers>
                 <VisualState.Setters>
                   <!-- Set Viewbox dimensions to FontSize -->
